add productId To createClientOrder
