+++ backend/src/Entity/OrderEntity.php
@@ -99,6 +99,11 @@ class OrderEntity
     private $clientID;
 
     /**
+     * @ORM\Column(type="integer", length=255, nullable=true)
+     */
+    private $productID;
+
+    /**
      * @Gedmo\Timestampable(on="create")
      * @ORM\Column(type="datetime", nullable=true)
      */
@@ -304,6 +309,18 @@ class OrderEntity
     public function setClientID(string $clientID): self
     {
         $this->clientID = $clientID;
+
+        return $this;
+    }
+
+    public function getProductID(): ?int
+    {
+        return $this->productID;
+    }
+
+    public function setProductID(int $productID): self
+    {
+        $this->productID = $productID;
 
         return $this;
     }
