--- backend/src/Entity/ProductOrderEntity.php
@@ -1,143 +1,0 @@
-<?php
-
-namespace App\Entity;
-
-use App\Repository\ProductOrderEntityRepository;
-use Doctrine\ORM\Mapping as ORM;
-
-/**
- * @ORM\Entity(repositoryClass=ProductOrderEntityRepository::class)
- */
-class ProductOrderEntity
-{
-    /**
-     * @ORM\Id
-     * @ORM\GeneratedValue
-     * @ORM\Column(type="integer")
-     */
-    private $id;
-
-    /**
-     * @ORM\Column(type="integer", nullable=true)
-     */
-    private $productID;
-
-    /**
-     * @ORM\Column(type="string", length=255, nullable=true)
-     */
-    private $clientID;
-
-    /**
-     * @ORM\Column(type="datetime", nullable=true)
-     */
-    private $createdAt;
-
-    /**
-     * @ORM\Column(type="string", length=255, nullable=true)
-     */
-    private $chatRoomId;
-
-    /**
-     * @ORM\Column(type="datetime", nullable=true)
-     */
-    private $updatedDate;
-
-    /**
-     * @ORM\Column(type="datetime", nullable=true)
-     */
-    private $deliveryDate;
-
-    /**
-     * @ORM\Column(type="string", length=255, nullable=true)
-     */
-    private $state;
-
-    public function getId(): ?int
-    {
-        return $this->id;
-    }
-
-    public function getProductID(): ?int
-    {
-        return $this->productID;
-    }
-
-    public function setProductID(?int $productID): self
-    {
-        $this->productID = $productID;
-
-        return $this;
-    }
-
-    public function getClientID(): ?string
-    {
-        return $this->clientID;
-    }
-
-    public function setClientID(?string $clientID): self
-    {
-        $this->clientID = $clientID;
-
-        return $this;
-    }
-
-    public function getCreatedAt(): ?\DateTimeInterface
-    {
-        return $this->createdAt;
-    }
-
-    public function setCreatedAt(?\DateTimeInterface $createdAt): self
-    {
-        $this->createdAt = $createdAt;
-
-        return $this;
-    }
-
-    public function getChatRoomId(): ?string
-    {
-        return $this->chatRoomId;
-    }
-
-    public function setChatRoomId(?string $chatRoomId): self
-    {
-        $this->chatRoomId = $chatRoomId;
-
-        return $this;
-    }
-
-    public function getUpdatedDate(): ?\DateTimeInterface
-    {
-        return $this->updatedDate;
-    }
-
-    public function setUpdatedDate(?\DateTimeInterface $updatedDate): self
-    {
-        $this->updatedDate = $updatedDate;
-
-        return $this;
-    }
-
-    public function getDeliveryDate(): ?\DateTimeInterface
-    {
-        return $this->deliveryDate;
-    }
-
-    public function setDeliveryDate(?\DateTimeInterface $deliveryDate): self
-    {
-        $this->deliveryDate = $deliveryDate;
-
-        return $this;
-    }
-
-    public function getState(): ?string
-    {
-        return $this->state;
-    }
-
-    public function setState(?string $state): self
-    {
-        $this->state = $state;
-
-        return $this;
-    }
-}
