--- backend/src/Repository/ProductOrderEntityRepository.php
@@ -1,50 +1,0 @@
-<?php
-
-namespace App\Repository;
-
-use App\Entity\ProductOrderEntity;
-use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
-use Doctrine\Persistence\ManagerRegistry;
-
-/**
- * @method ProductOrderEntity|null find($id, $lockMode = null, $lockVersion = null)
- * @method ProductOrderEntity|null findOneBy(array $criteria, array $orderBy = null)
- * @method ProductOrderEntity[]    findAll()
- * @method ProductOrderEntity[]    findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
- */
-class ProductOrderEntityRepository extends ServiceEntityRepository
-{
-    public function __construct(ManagerRegistry $registry)
-    {
-        parent::__construct($registry, ProductOrderEntity::class);
-    }
-
-    // /**
-    //  * @return ProductOrderEntity[] Returns an array of ProductOrderEntity objects
-    //  */
-    /*
-    public function findByExampleField($value)
-    {
-        return $this->createQueryBuilder('p')
-            ->andWhere('p.exampleField = :val')
-            ->setParameter('val', $value)
-            ->orderBy('p.id', 'ASC')
-            ->setMaxResults(10)
-            ->getQuery()
-            ->getResult()
-        ;
-    }
-    */
-
-    /*
-    public function findOneBySomeField($value): ?ProductOrderEntity
-    {
-        return $this->createQueryBuilder('p')
-            ->andWhere('p.exampleField = :val')
-            ->setParameter('val', $value)
-            ->getQuery()
-            ->getOneOrNullResult()
-        ;
-    }
-    */
-}
